Slash commands and stringbuilds now each have their own class and file
